+++ bot/loaders/__init__.py
@@ -1,2 +1,3 @@
+from .singlefile_html import load_singlefile_html
 from .video_transcript import load_video_transcript
 from .youtube_transcript import load_youtube_transcript
